update main.py
update main with minor changes. New way of starting the program from the command line. Now using argparser, which makes the process easier. Disabled physic for graph. Updated the way the output is looking 

Reminder: currently it is possible to enter a sourcefile that doesn`t exists and a non boolean value for the browser